pyproject.toml
[project]
name = "rag-with-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "cohere>=5.15.0",
    "fastapi>=0.115.12",
    "google-genai>=1.14.0",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
    "pydantic>=2.11.4",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.14.2",
    "uvicorn>=0.34.2",
]
